JAVA-1263: Move Compressors into own files, separate direct & heap code
- Move compressors into their own source files
- Separate out direct and heap compression code
- Change LZ4Compressor to use byte[] if not direct.  This is needed
  until JAVA-1306 is implemented.
- Use inputNioBuffer and outputNioBuffer for having a standard way of
  accessing internalNioBuffer (if possible).
- Add isolated tests for heap compression path.
