+++ clirr-ignores.xml
@@ -56,4 +56,16 @@
         <justification>This class is only present if the project was compiled with JDK 8+</justification>
     </difference>
 
+    <difference>
+        <differenceType>8001</differenceType> <!-- class removed -->
+        <className>com/datastax/driver/core/FrameCompressor$SnappyCompressor</className>
+        <justification>False positive, the enclosing class is package-private so this was never exposed</justification>
+    </difference>
+
+    <difference>
+        <differenceType>8001</differenceType> <!-- class removed -->
+        <className>com/datastax/driver/core/FrameCompressor$LZ4Compressor</className>
+        <justification>False positive, the enclosing class is package-private so this was never exposed</justification>
+    </difference>
+
 </differences>
